no message
git-svn-id: svn://localhost/trunk@793 3bdad9f7-c964-442a-868d-553b97b55ca0

+++ surfit/src/surfit/f_completer.cpp
@@ -203,7 +203,7 @@ bool f_completer::minimize() {
 			}
 		}
 
-		int matrix_size = method_basis_cntX*method_basis_cntY;
+		size_t matrix_size = method_basis_cntX*method_basis_cntY;
 
 		std::vector<short int> * flood_areas = new std::vector<short int>(matrix_size);
 		int flood_areas_cnt = 0;
